pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "marketing-openenv"
version = "0.1.0"
description = "Simulates a real-world marketing team workflow where an agent must coordinate tools to create publishable content."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv",
    "fastapi",
    "pydantic",
    "google-generativeai",
    "uvicorn",
    "python-dotenv",
    "httpx",
    "jinja2",
    "python-multipart"
]

[project.scripts]
marketing-openenv-server = "server.app:start"
